+++ rpcs3/rpcs3qt/save_data_list_dialog.cpp
@@ -239,7 +239,7 @@ void save_data_list_dialog::LoadEntries(void)
 void save_data_list_dialog::UpdateList(void)
 {
 	m_list->clearContents();
-	m_list->setRowCount(m_save_entries.size());
+	m_list->setRowCount((int)m_save_entries.size());
 	gui_settings settings(this);
 
 	int row = 0;
